fix: move plans to dedicated page
